MNT: Rename ui_redesign_test.py -> main.py
