+++ trace/main.py
@@ -2,82 +2,232 @@ import os
 import argparse
 import subprocess
 from socket import gethostname
-from typing import Dict, List, Tuple, Union
 from getpass import getuser
-from logging import Handler, LogRecord
 from datetime import datetime
 
-import pyqtgraph.exporters
-from qtpy.sip import isdeleted
-from qtpy.QtCore import Qt, Slot
-from qtpy.QtWidgets import QLabel, QFileDialog, QApplication, QAbstractButton
+import qtawesome as qta
+from qtpy.QtGui import QFont, QImage, QKeySequence
+from qtpy.QtCore import Qt, Slot, QSize, Signal, QBuffer, QIODevice
+from qtpy.QtWidgets import (
+    QMenu,
+    QLabel,
+    QDialog,
+    QWidget,
+    QMenuBar,
+    QSplitter,
+    QFileDialog,
+    QHBoxLayout,
+    QMessageBox,
+    QPushButton,
+    QSizePolicy,
+    QSpacerItem,
+    QVBoxLayout,
+    QApplication,
+    QButtonGroup,
+)
+from pyqtgraph.exporters import ImageExporter
+from services.elog_client import get_user, post_entry
 
 from pydm import Display
+from pydm.widgets import PyDMLabel, PyDMArchiverTimePlot
 from pydm.utilities.macro import parse_macro_string
 
 from config import logger, datetime_pv
-from mixins import FileIOMixin, AxisTableMixin, PlotConfigMixin, TracesTableMixin
-from styles import CenterCheckStyle
-from widgets import DataInsightTool
-from trace_file_convert import PathAction
+from file_io import PathAction, TraceFileHandler
+from widgets import ControlPanel, DataInsightTool, PlotSettingsModal
+from widgets.elog_post_modal import ElogPostModal
+
+DISABLE_AUTO_SCROLL = -2  # Using -2 as invalid since QButtonGroups use -1 as invalid
 
 
-class TraceDisplay(Display, TracesTableMixin, AxisTableMixin, FileIOMixin, PlotConfigMixin):
-    def __init__(self, parent=None, args=None, macros=None, ui_filename=__file__.replace(".py", ".ui")) -> None:
-        super(TraceDisplay, self).__init__(parent=parent, args=args, macros=macros, ui_filename=ui_filename)
-        # Set up PyDMApplication
+class TraceDisplay(Display):
+    gridline_opacity_change = Signal(int)
+    set_all_y_axis_gridlines = Signal(bool)
+
+    def __init__(self, parent=None, args=None, macros=None) -> None:
+        super(TraceDisplay, self).__init__(parent=parent, args=args, macros=macros, ui_filename=None)
+        self.build_ui()
         self.configure_app()
-        self.set_footer()
+        self.resize(1000, 600)
 
-        # Initialize the Mixins
-        self.axis_table_init()
-        self.traces_table_init()
-        self.plot_config_init()
-        self.file_io_init()
+        # Set plot's timerange after the UI is built
+        default_button = self.timespan_buttons.button(3600)
+        default_button.setChecked(True)
 
-        self.curve_delegates_init()
-        self.axis_delegates_init()
-        self.timespan = -1
-        self.axis_table_model.reset_everything.connect(self.resetPlot)
-        # Create reference dict for timespan_btns button group
-        self.button_spans = {
-            self.ui.min_scale_btn: 60,
-            self.ui.hour_scale_btn: 3600,
-            self.ui.day_scale_btn: 86400,
-            self.ui.week_scale_btn: 604800,
-            self.ui.month_scale_btn: 2628300,
-            self.ui.cursor_scale_btn: -1,
-        }
-        self.ui.timespan_btns.buttonToggled.connect(self.set_plot_timerange)
-
-        self.ui.dit_btn.clicked.connect(self.open_data_insight_tool)
-        self.ui.save_img_btn.clicked.connect(self.save_plot_image)
-        self.ui.fetch_archive_btn.clicked.connect(self.fetch_archive)
-
-        # Toggle "Cursor" button on plot-mouse interaction
-        multi_axis_plot = self.ui.main_plot.plotItem
-        multi_axis_plot.vb.menu = None
-        multi_axis_plot.sigXRangeChangedManually.connect(self.ui.cursor_scale_btn.toggle)
-
-        # Additional configuration of main plot
-        self.ui.main_plot.cache_data = False
-        self.ui.main_plot._show_all = False
-        self.ui.main_plot.show_extension_lines = True
-
-        # Parse macros & arguments, then include them in startup
-        input_file, startup_pvs = self.parse_macros_and_args(macros, args)
+        input_file, startup_pvs = self.parse_cli_args(args, macros)
         if input_file:
-            self.import_save_file(input_file)
+            self.file_handler.open_file(input_file)
         for pv in startup_pvs:
-            if pv in self.curves_model:
-                continue
-            last_row = self.curves_model.rowCount() - 1
-            index = self.curves_model.index(last_row, 0)
-            self.curves_model.setData(index, pv, Qt.EditRole)
+            self.layout().itemAt(0).widget().widget(1).add_curve(pv)
 
-    def menu_items(self) -> dict:
-        """Add export & import functionality to File menu"""
-        return {"Export": (self.export_save_file, "Ctrl+S"), "Import": (self.import_save_file, "Ctrl+L")}
+    @property
+    def gridline_opacity(self) -> int:
+        """Get the current gridline opacity value from the plot settings"""
+        return self.plot_settings.gridline_opacity
+
+    def minimumSizeHint(self):
+        return QSize(700, 350)
+
+    def build_ui(self) -> None:
+        # Set window title
+        self.setWindowTitle("Trace")
+        # Create main layout
+        main_layout = QVBoxLayout()
+        self.setLayout(main_layout)
+
+        # Create the plotting and control widgets
+        plot_side_widget = self.build_plot_side(self)
+        self.control_panel = ControlPanel()
+        self.control_panel.layout().setContentsMargins(8, 0, 0, 0)
+        self.control_panel.plot = self.plot
+        self.control_panel.curve_list_changed.connect(self.data_insight_tool.update_pv_select_box)
+
+        # Create main splitter
+        main_splitter = QSplitter(self)
+        main_splitter.addWidget(plot_side_widget)
+        main_splitter.addWidget(self.control_panel)
+        main_splitter.setCollapsible(0, False)
+        main_splitter.setStretchFactor(0, 1)
+        main_splitter.setHandleWidth(10)
+        main_splitter.setStyleSheet(
+            "\n".join(
+                [
+                    "QSplitter::handle {",
+                    "  background-color: white;",
+                    "}",
+                ]
+            )
+        )
+        main_layout.addWidget(main_splitter)
+
+        # Create the footer section of the app
+        footer_widget = self.build_footer(self)
+        main_layout.addWidget(footer_widget)
+
+    def build_plot_side(self, parent):
+        plot_side_widget = QWidget(parent)
+        plot_side_layout = QVBoxLayout()
+        plot_side_layout.setContentsMargins(0, 0, 8, 0)
+        plot_side_widget.setLayout(plot_side_layout)
+
+        toolbar = self.build_toolbar(plot_side_widget)
+        plot_side_layout.addWidget(toolbar)
+
+        # Create plot
+        self.plot = PyDMArchiverTimePlot(
+            plot_side_widget,
+            background="white",
+            optimized_data_bins=5000,
+            cache_data=False,
+            show_all=False,
+        )
+
+        multi_axis_plot = self.plot.plotItem
+        multi_axis_plot.vb.menu = None
+        multi_axis_plot.sigXRangeChangedManually.connect(self.disable_auto_scroll_button.click)
+        plot_side_layout.addWidget(self.plot)
+
+        self.data_insight_tool = DataInsightTool(self)
+        self.data_insight_tool.plot = self.plot
+
+        self.settings_button = QPushButton(self.plot)
+        self.settings_button.setIcon(qta.icon("msc.settings-gear"))
+        self.settings_button.setFlat(True)
+
+        self.plot_settings = PlotSettingsModal(self.settings_button, self.plot)
+        self.plot_settings.auto_scroll_interval_change.connect(self.set_auto_scroll_interval)
+        self.plot_settings.grid_alpha_change.connect(self.gridline_opacity_change.emit)
+        self.plot_settings.set_all_y_axis_gridlines.connect(self.plot.setShowYGrid)
+        self.plot_settings.set_all_y_axis_gridlines.connect(self.set_all_y_axis_gridlines.emit)
+        self.plot_settings.disable_autoscroll.connect(self.disable_auto_scroll_button.click)
+        self.settings_button.clicked.connect(self.plot_settings.show)
+
+        return plot_side_widget
+
+    def build_toolbar(self, parent):
+        toolbar_widget = QWidget(parent)
+        # Create tool layout
+        tool_layout = QHBoxLayout()
+        tool_layout.setContentsMargins(0, 0, 0, 0)
+        toolbar_widget.setLayout(tool_layout)
+
+        tool_spacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
+        tool_layout.addSpacerItem(tool_spacer)
+
+        timespan_buttons = self.build_timespan_buttons(toolbar_widget)
+        tool_layout.addWidget(timespan_buttons)
+
+        return toolbar_widget
+
+    def build_timespan_buttons(self, parent: QWidget):
+        timespan_button_widget = QWidget(parent)
+        timespan_button_layout = QHBoxLayout()
+        timespan_button_layout.setContentsMargins(0, 0, 0, 0)
+        timespan_button_widget.setLayout(timespan_button_layout)
+
+        self.timespan_buttons = QButtonGroup(timespan_button_widget)
+        self.timespan_buttons.setExclusive(True)
+
+        timespan_button_data = (
+            ("1m", 60),
+            ("1h", 3600),
+            ("1d", 86400),
+            ("1w", 604800),
+            ("1M", 2628300),
+            ("Disable AutoScroll", DISABLE_AUTO_SCROLL),
+        )
+
+        for text, id in timespan_button_data:
+            timespan_button = QPushButton(text, timespan_button_widget)
+            timespan_button.setMaximumWidth(35)
+            timespan_button.setCheckable(True)
+            timespan_button_layout.addWidget(timespan_button)
+            self.timespan_buttons.addButton(timespan_button, id)
+
+        self.disable_auto_scroll_button = self.timespan_buttons.button(DISABLE_AUTO_SCROLL)
+        self.disable_auto_scroll_button.hide()
+
+        self.timespan_buttons.buttonToggled.connect(self.set_auto_scroll_span)
+
+        return timespan_button_widget
+
+    def build_footer(self, parent: QWidget):
+        label_font = QFont()
+        label_font.setPointSize(8)
+
+        footer_widget = QWidget(parent)
+        footer_widget.setFixedHeight(12)
+        footer_layout = QHBoxLayout()
+        footer_layout.setContentsMargins(0, 0, 0, 0)
+        footer_widget.setLayout(footer_layout)
+
+        footer_label_data = (
+            (self.git_version(), "Trace Version"),
+            (gethostname(), "Node Name"),
+            (getuser(), "User Name"),
+            (str(os.getpid()), "PID"),
+            (os.getenv("PYDM_ARCHIVER_URL"), "Archiver URL"),
+        )
+
+        for text, tooltip in footer_label_data:
+            label = QLabel(text, footer_widget)
+            label.setFont(label_font)
+            label.setToolTip(tooltip)
+            label.setAlignment(Qt.AlignBottom)
+            footer_layout.addWidget(label)
+            footer_layout.addWidget(BreakerLabel(footer_widget))
+
+        last_breaker = footer_widget.children()[-1]
+        footer_layout.removeWidget(last_breaker)
+
+        footer_spacer = QSpacerItem(40, 12, QSizePolicy.Expanding, QSizePolicy.Minimum)
+        footer_layout.addSpacerItem(footer_spacer)
+
+        self.time_label = PyDMLabel(footer_widget, f"ca://{datetime_pv}")
+        self.time_label.setAlignment(Qt.AlignBottom)
+        footer_layout.addWidget(self.time_label)
+
+        return footer_widget
 
     def configure_app(self):
         """UI changes to be made to the PyDMApplication"""
@@ -93,137 +243,54 @@ class TraceDisplay(Display, TracesTableMixin, AxisTableMixin, FileIOMixin, PlotC
         app.main_window.toggle_status_bar(False)
         app.main_window.ui.actionShow_Status_Bar.setChecked(False)
 
-        # Add style to center checkboxes in table cells
-        app.setStyle(CenterCheckStyle())
+        # Create a TraceFileController instance for handling file I/O operations
+        self.file_handler = TraceFileHandler(self.plot, self)
+        self.file_handler.axes_signal.connect(self.control_panel.set_axes)
+        self.file_handler.curves_signal.connect(self.control_panel.set_curves)
+        self.file_handler.plot_settings_signal.connect(self.plot_settings.plot_setup)
+        self.file_handler.auto_scroll_span_signal.connect(self.set_auto_scroll_span)
+        self.file_handler.timerange_signal.connect(self.set_plot_timerange)
 
-        # Adjust settings for main_spltr
-        self.ui.main_spltr.setSizes([1, 200])
-        self.ui.main_spltr.setCollapsible(0, False)
-        self.ui.main_spltr.setStretchFactor(0, 1)
+        # Remove shortcut from the "Open File" menu action
+        open_file_action = app.main_window.ui.actionOpen_File
+        open_file_action.setText("Open PyDM File...")
+        open_file_action.setShortcut(QKeySequence())
 
-    def set_footer(self):
-        """Set footer information for application. Includes logging, nodename,
-        username, PID, git version, Archiver URL, and current datetime
-        """
-        self.logging_handler = LoggingHandler(self.ui.logger_lbl)
-        logger.addHandler(self.logging_handler)
-        logger.setLevel("NOTSET")
+        # Create a custom menu for the application
+        menu_bar: QMenuBar = app.main_window.ui.menubar
+        first_menu = app.main_window.ui.menuFile.menuAction()
+        trace_menu = self.construct_trace_menu(menu_bar)
+        menu_bar.insertMenu(first_menu, trace_menu)
 
-        self.ui.ftr_ver_lbl.setText(self.git_version())
-        self.ui.ftr_node_lbl.setText(gethostname())
-        self.ui.ftr_user_lbl.setText(getuser())
-        self.ui.ftr_pid_lbl.setText(str(os.getpid()))
-        self.ui.ftr_url_lbl.setText(os.getenv("PYDM_ARCHIVER_URL"))
-        self.ui.ftr_time_lbl.channel = "ca://" + datetime_pv
+    def construct_trace_menu(self, parent: QMenuBar) -> QMenu:
+        """Create the menu for the application."""
+        menu = QMenu("Trace", parent)
+        save = menu.addAction("Save", self.file_handler.save_file)
+        save.setShortcut(QKeySequence("Ctrl+S"))
+        save_as = menu.addAction("Save As...", self.file_handler.save_as)
+        save_as.setShortcut(QKeySequence("Ctrl+Shift+S"))
+        load = menu.addAction("Open Trace Config...", self.file_handler.open_file)
+        load.setShortcut(QKeySequence("Ctrl+O"))
+        menu.addSeparator()
 
-    def parse_macros_and_args(self, macros: Dict[str, Union[str, list]], args: List[str]) -> Tuple[str, list]:
-        """Parse user provided macros and args into lists of PVs to use on
-        startup or which file to import on startup
+        save_image = menu.addAction("Save Plot Image...", self.save_plot_image)
+        save_image.setShortcut(QKeySequence("Ctrl+I"))
+        save_elog = menu.addAction("Save ELOG Entry...", self.elog_button_clicked)
+        save_elog.setShortcut(QKeySequence("Ctrl+E"))
+        menu.addSeparator()
 
-        Parameters
-        ----------
-        macros : Dict[str, str | list]
-            Dictionary containing all of the macros passed into PyDM
-        args : List[str]
-            List of all arguments passed into the application to be parsed
+        fetch_archive = menu.addAction("Fetch Archive Data", self.fetch_archive)
+        fetch_archive.setShortcut(QKeySequence("Ctrl+F"))
+        dit_action = menu.addAction("Data Insight Tool...", self.data_insight_tool.show)
+        dit_action.setShortcut(QKeySequence("Ctrl+D"))
 
-        Returns
-        -------
-        tuple
-            A tuple containing the file to import from and the list of PVs to use on startup
-        """
-        # Default macros is None
-        if not macros:
-            macros = {}
+        return menu
 
-        # Construct an argument parser for args
-        trace_parser = argparse.ArgumentParser(
-            prog="trace",
-            description="Trace\nThis is a PyDM application used to display archived and live pv data.",
-            epilog="Examples:\n\tpydm $PHYSICS_TOP/trace/main.py"
-            + "\n\tbash $PHYSICS_TOP/trace/launch_trace.bash"
-            + "\n\t%(prog)s"
-            + "\n\t%(prog)s -i some_input_file.trc"
-            + "\n\t%(prog)s -p SOME:PV:TO:PLOT OTHER:PV:TO:PLOT"
-            + '\n\t%(prog)s -m \'{"PVS": ["FOO:CHANNEL", "BAR:CHANNEL", "f://{A}+{B}"]}\''
-            + '\n\t%(prog)s -m "INPUT_FILE = trace/examples/FormulaExample.trc"',
-            formatter_class=argparse.RawTextHelpFormatter,
-        )
-        trace_parser.add_argument("-V", "--version", action="version", version="%(prog)s " + self.git_version())
-        trace_parser.add_argument(
-            "-i",
-            "--input_file",
-            action=PathAction,
-            nargs="?",
-            default=[],
-            help="Absolute file path to import from\nAlternatively can be provided as INPUT_FILE macro",
-        )
-        trace_parser.add_argument(
-            "-p",
-            "--pvs",
-            nargs="*",
-            default=[],
-            help="Space-separated list of PVs to show on startup\nFormulas should be passed without spaces: "
-            + "f://{A}+{B}\nAlternatively can be provided as PV or PVS macros",
-        )
-        trace_parser.add_argument(
-            "-m",
-            "--macro",
-            default="",
-            help="Mimic PyDM macro replacements to use. Should be in JSON object format."
-            + "\nJSON Formatting Reminder:"
-            + "\n\tJSON requires double quotes for strings, so you should wrap this"
-            + "\n\twhole argument in single quotes.\n\t--or--"
-            + "\n\tSpecify macro replacements as KEY=value pairs using a comma as a"
-            + "\n\tdelimiter. If you want to uses spaces after the delimiters or around"
-            + "\n\tthe '=' signs, wrap the entire set with quotes.",
-        )
-
-        # Parse arguments and ignore unknowns
-        trace_args, unknown = trace_parser.parse_known_args(args)
-        for u in unknown:
-            if not u:
-                continue
-            logger.warning(f"Not using unknown arguments: {u}")
-
-        # Parse any macros passed into trace
-        if trace_args.macro:
-            parsed_macros = parse_macro_string(trace_args.macro)
-            macros.update(**parsed_macros)
-
-        # Get the file to import from if one is provided. Prioritize args over macro
-        input_file = ""
-        if trace_args.input_file:
-            # Need to unpack as PathAction returns a list
-            input_file = trace_args.input_file[0]
-        elif "INPUT_FILE" in macros:
-            input_file = macros["INPUT_FILE"]
-
-        # Get the list of PVs to show on startup
-        startup_pvs = []
-        for key in ("PV", "PVS"):
-            if key in macros:
-                val = macros[key]
-                if isinstance(val, str):
-                    startup_pvs.append(val)
-                elif isinstance(val, list):
-                    startup_pvs += val
-        startup_pvs += trace_args.pvs
-
-        # Remove duplicates from startup_pvs
-        startup_pvs = list(dict.fromkeys(startup_pvs))
-
-        return (input_file, startup_pvs)
-
-    def open_data_insight_tool(self):
-        """Create a new instance of the Data Insight Tool"""
-        dit = DataInsightTool(self, self.curves_model, self.ui.main_plot)
-        dit.show()
-
+    @Slot()
     def save_plot_image(self) -> None:
         """Saves current plot as an image. Opens file dialog to allow user to
         set custom location."""
-        exporter = pyqtgraph.exporters.ImageExporter(self.ui.main_plot.plotItem)
+        exporter = ImageExporter(self.plot.plotItem)
         default_filename = datetime.now().strftime(f"{getuser()}_trace_%Y%m%d_%H%M%S.png")
         usr_home_dir = os.path.expanduser("~")
         file_path, _ = QFileDialog.getSaveFileName(
@@ -239,13 +306,122 @@ class TraceDisplay(Display, TracesTableMixin, AxisTableMixin, FileIOMixin, PlotC
             except Exception as e:
                 logger.error(f"Failed to save image: {e}")
 
+    @Slot()
+    def elog_button_clicked(self) -> bool:
+        """Takes a snapshot of the plot and posts it to the Elog API.
+
+        :return: True if the post was successful, False otherwise."""
+        # Test if API is reachable
+        status_code, _ = get_user()
+        if status_code != 200:
+            error_dialog = QMessageBox()
+            error_dialog.setIcon(QMessageBox.Warning)
+            error_dialog.setWindowTitle("Connection Error")
+            error_dialog.setText("Failed to connect to the Elog API.")
+            error_dialog.setInformativeText(
+                f"""No entry was posted. If this issue persists, please report it in the
+                #elog-general Slack channel. \n\nError Code: {status_code}"""
+            )
+            error_dialog.setStandardButtons(QMessageBox.Ok)
+            error_dialog.exec_()
+            return False
+
+        # Form the request info
+        # Use ImageExporter to take a snapshot of the plot
+        exporter = ImageExporter(self.plot.plotItem)
+        img: QImage = exporter.export(toBytes=True)
+        # Convert Qimage to bytes
+        buffer = QBuffer()
+        buffer.open(QIODevice.ReadWrite)
+        img.save(buffer, "PNG")
+        image_bytes = buffer.data()
+        # Get entry info from user
+        dialog = ElogPostModal.maybe_create(self, image_bytes=image_bytes)
+        if dialog is not None and dialog.exec_() == QDialog.Accepted:
+            title, body, logbooks = dialog.get_inputs()
+        else:
+            return False
+
+        # Post the request to the Elog API
+        status_code, _ = post_entry(title, body, logbooks, image_bytes)
+
+        # Check if the request was successful
+        if status_code == 201:
+            success_dialog = QMessageBox()
+            success_dialog.setIcon(QMessageBox.Information)
+            success_dialog.setWindowTitle("Elog Entry Posted")
+            success_dialog.setText("Elog entry posted successfully!")
+            success_dialog.setStandardButtons(QMessageBox.Ok)
+            success_dialog.exec_()
+            return True
+        else:
+            error_dialog = QMessageBox()
+            error_dialog.setIcon(QMessageBox.Warning)
+            error_dialog.setWindowTitle("Connection Error")
+            error_dialog.setText("Failed to connect to the Elog API.")
+            error_dialog.setInformativeText(
+                f"No entry was posted. If this issue persists, please report it in the \
+                #elog-general Slack channel. \n\nError Code: {status_code}"
+            )
+            error_dialog.setStandardButtons(QMessageBox.Ok)
+            error_dialog.exec_()
+            return False
+
+    @Slot()
     def fetch_archive(self) -> None:
         """Triggers a fetch to the archive"""
-        if not (self.ui.main_plot._archive_request_queued):
+        if not (self.plot._archive_request_queued):
             logger.info("Requesting data from archiver")
-            self.ui.main_plot.requestDataFromArchiver()
+            self.plot.requestDataFromArchiver()
         else:
             logger.info("Archive fetch is already queued")
+
+    @Slot(tuple)
+    def set_plot_timerange(self, timerange: tuple[float, float]) -> None:
+        """Set the plot's timerange to the given start and end datetimes."""
+        self.disable_auto_scroll_button.click()
+        self.plot.setXRange(*timerange)
+        logger.debug(f"Plot timerange set to {timerange[0]} - {timerange[1]}")
+
+    @Slot()
+    @Slot(float)
+    def set_auto_scroll_span(self, timespan: float = None) -> None:
+        """Slot to be called when a timespan setting button is pressed.
+        This will enable autoscrolling along the x-axis and disable mouse
+        controls. If the "Cursor" button is pressed, then autoscrolling is
+        disabled and mouse controls are enabled.
+        """
+        if timespan is None:
+            timespan = self.timespan_buttons.checkedId()
+            enable_scroll = timespan != DISABLE_AUTO_SCROLL
+        else:
+            enable_scroll = True
+            self.disable_auto_scroll_button.click()
+
+        if enable_scroll:
+            logger.debug(f"Enabling plot autoscroll for {timespan}s")
+        else:
+            logger.debug("Disabling plot autoscroll, using mouse controls")
+        self.autoScroll(enable=enable_scroll, timespan=timespan)
+
+    @Slot(int)
+    def set_auto_scroll_interval(self, inteval: int) -> None:
+        """Set the auto scroll interval for the plot"""
+        timespan = self.timespan_buttons.checkedId()
+        enable_scroll = timespan != DISABLE_AUTO_SCROLL
+
+        self.plot.setAutoScroll(enable_scroll, timespan, refresh_rate=inteval)
+
+    @Slot(bool)
+    @Slot(bool, float)
+    def autoScroll(self, enable: bool, timespan: float = None):
+        if timespan is None:
+            timespan = self.timespan_buttons.checkedId()
+            if timespan < 0:
+                return
+
+        refresh_interval = self.plot_settings.auto_scroll_interval
+        self.plot.setAutoScroll(enable, timespan, refresh_rate=refresh_interval)
 
     @staticmethod
     def git_version():
@@ -256,61 +432,111 @@ class TraceDisplay(Display, TracesTableMixin, AxisTableMixin, FileIOMixin, PlotC
         )
         return git_cmd.stdout.strip()
 
-    @Slot()
-    def resetPlot(self) -> None:
-        """Reset the Axis model and the Curve model to empty states"""
-        self.axis_table_model.set_model_axes()
-        self.curves_model.set_model_curves()
+    def parse_cli_args(self, args, macros):
+        """"""
+        args = args or []
+        macros = macros or {}
 
-    @Slot(QAbstractButton, bool)
-    def set_plot_timerange(self, button: QAbstractButton, toggled: bool) -> None:
-        """Slot to be called when a timespan setting button is pressed.
-        This will enable autoscrolling along the x-axis and disable mouse
-        controls. If the "Cursor" button is pressed, then autoscrolling is
-        disabled and mouse controls are enabled.
+        parser = argparse.ArgumentParser(
+            prog="trace",
+            description="Trace\nThis is a PyDM application used to display archived and live pv data.",
+            epilog="\n\t".join(
+                [
+                    "Examples:",
+                    "pydm $PHYSICS_TOP/trace/main.py"
+                    "bash $PHYSICS_TOP/trace/launch_trace.bash"
+                    "%(prog)s"
+                    "%(prog)s -i some_input_file.trc"
+                    "%(prog)s -p SOME:PV:TO:PLOT OTHER:PV:TO:PLOT"
+                    '%(prog)s -m \'{"PVS": ["FOO:CHANNEL", "BAR:CHANNEL", "f://{A}+{B}"]}\''
+                    '%(prog)s -m "INPUT_FILE = trace/examples/FormulaExample.trc"',
+                ]
+            ),
+            formatter_class=argparse.RawTextHelpFormatter,
+        )
 
-        Parameters
-        ----------
-        button : QAbstractButton
-            The timespan setting button pressed. Determines which timespan
-            to set.
-        toggled : bool
-            Whether or not the associated button is toggled or not.
-        """
-        if not toggled:
-            return
+        parser.add_argument("-v", "--version", action="version", version="%(prog)s " + self.git_version())
+        parser.add_argument(
+            "-i",
+            "--input_file",
+            action=PathAction,
+            nargs="?",
+            default=[],
+            help="Absolute file path to import from\nAlternatively can be provided as INPUT_FILE macro",
+        )
+        parser.add_argument(
+            "-p",
+            "--pvs",
+            nargs="*",
+            default=[],
+            help="\n".join(
+                [
+                    "Space-separated list of PVs to show on startup",
+                    "Formulas should be passed without spaces: f://{A}+{B}",
+                    "Alternatively can be provided as PV or PVS macros",
+                ]
+            ),
+        )
+        parser.add_argument(
+            "-m",
+            "--macro",
+            default="",
+            help="\n\t".join(
+                [
+                    "Mimic PyDM macro replacements to use. Should be in JSON object format.",
+                    "ON Formatting Reminder:",
+                    "JSON requires double quotes for strings, so you should wrap this",
+                    "whole argument in single quotes.",
+                    "--or--",
+                    "Specify macro replacements as KEY=value pairs using a comma as a",
+                    "delimiter. If you want to uses spaces after the delimiters or around",
+                    "the '=' signs, wrap the entire set with quotes.",
+                ]
+            ),
+        )
 
-        logger.debug("Setting plot timerange")
-        if button not in self.button_spans:
-            logger.error(f"{button} is not a valid timespan button")
-            return
-        enable_scroll = button != self.ui.cursor_scale_btn
-        self.timespan = self.button_spans[button]
-        if enable_scroll:
-            logger.debug(f"Enabling plot autoscroll for {self.timespan}s")
-        else:
-            logger.debug("Disabling plot autoscroll, using mouse controls")
-        self.autoScroll(enable=enable_scroll)
+        # Parse arguments and ignore unknowns
+        known, unknown = parser.parse_known_args(args)
+        for arg in unknown:
+            if arg:
+                logger.warning(f"Not using unknown argument: {arg}")
+
+        # Parse any macros passed into trace
+        if known.macro:
+            parsed_macros = parse_macro_string(known.macro)
+            macros.update(**parsed_macros)
+
+        # Get the file to import from if one is provided. Prioritize args over macro
+        try:
+            # Need to unpack as PathAction returns a list
+            input_file = known.input_file[0]
+        except IndexError:
+            input_file = macros.get("INPUT_FILE", "")
+
+        # Get the list of PVs to show on startup
+        startup_pvs = []
+        for key in ("PV", "PVS"):
+            if key in macros:
+                val = macros[key]
+                if isinstance(val, str):
+                    startup_pvs.append(val)
+                elif isinstance(val, list):
+                    startup_pvs.extend(val)
+        startup_pvs += known.pvs
+
+        # Remove duplicates from startup_pvs
+        startup_pvs = list(dict.fromkeys(startup_pvs))
+
+        return (input_file, startup_pvs)
 
 
-class LoggingHandler(Handler):
-    def __init__(self, logging_lbl: QLabel, level: int = 0) -> None:
-        super().__init__(level)
-        self.logging_lbl = logging_lbl
+class BreakerLabel(QLabel):
+    breaker_font = QFont()
+    breaker_font.setBold(True)
+    breaker_font.setPointSize(12)
 
-    def emit(self, record: LogRecord):
-        """Any logs from the logger will be displayed on the logging label. If
-        the log level is greater than 20 (INFO), then the level will be shown as
-        well. Also checks if the logging label has been deleted.
-
-        Parameters
-        ----------
-        record : LogRecord
-            The logger's log record.
-        """
-        if isdeleted(self.logging_lbl):
-            return
-        log = record.msg
-        if record.levelno > 20:
-            log = f"[{record.levelname}] - {log}"
-        self.logging_lbl.setText(log)
+    def __init__(self, parent):
+        super().__init__(parent)
+        self.setText("|")
+        self.setFont(self.breaker_font)
+        self.setAlignment(Qt.AlignBottom)
